+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 
 - [Integration](https://crates.io/crates/mcan-core) with the [`mcan`](https://crates.io/crates/mcan) crate.
 - Implementation of blocking::i2c::Transactional trait from [embedded-hal](https://crates.io/crates/embedded-hal) for TWI device.
+- Implementation of ReadNorFlash and NorFlash traits from [embedded-storage](https://crates.io/crates/embedded-storage) for EFC.
 
 ### Changed
 - Remove `rust-toolchain.toml` and control MSRV from .github/workflow/ files instead.